Added DrawerMenu - installed and implemented the drawerMenu from react-navigation - implemented the Authentication Flow - added new assets from Luke's designs (via Sketch) - added Wallet, Transactions, Scan and Send & Receive views mocks
