+++ Router.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import { DrawerNavigator } from 'react-navigation';
+
+// The DrawerNavigator uses all the screens
+
+export const MyDrawer = DrawerNavigator({
+});
